+++ CHANGES.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     * add bls curves types (`bls12_381_fr`, `bls12_381_g1` and `bls12_381_g2`), arthmetic (`+`, `*`, unary `-`) and operation (`pairing_check`)
   - `update` function for map (like `UPDATE` in michelson)
   - `state is` section in entry
+  - `called by` supports asset with a key typed address
 
 ### Changed
   - `slice` returns `option<byte|string>`